Test with a true cluster
Test with a true cluster where each node is a different suite and use
its own docker. For that, we need dnsdock to resolve node name.
dnsdock is provided also in a docker in a different suite.

To make this work, we need to monkey patch the way suites are
scheduled.

Add serverspec configuration to check if the zookeeper cluster is
really running. Test all hosts. Import zookeeper gem.