Use a special docker image, switch to docker_cli
After the release of centos 7.2, many problems came from running a
docker with systemd in privileged mode. By chance, it became also
quite easy to run it in normal mode. However, the limitations in
the customization of images provided by kitchen-docker, added to
the increasing burden of rationalizing them between cookbooks,
made a good case for using a special image hosted on docker hub.

This is why I created the image sbernard/centos-systemd-kitchen,
configured to run systemd as init, with chef verifiers preinstalled,
making it a perfect base to run kitchen tests. This patch uses it.

Also, it changes the kitchen driver from docker to docker_cli, which
is simpler, better maintained and more in docker spirit: it
communicates with docker by 'exec' instead of ssh.