Allow to not install java package
Users may want to install java via other ways, and current
systemd_recipe does not support non-installation of java by setting the
java_package to '' as the chef package resource requires a name or it
will trigger an error.

This commit only installs the java when users define the package for
the node platform, otherwise it outputs a WARN message.